+++ test/test/algorithm/EuclideanDistanceTest.java
@@ -1,4 +1,4 @@
-package test;
+package test.algorithm;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
