+++ test/test/io/ParserTest.java
@@ -1,4 +1,4 @@
-package test;
+package test.io;
 
 import java.io.IOException;
 import java.util.List;
