+++ test/test/mapper/LocationMapperTest.java
@@ -1,4 +1,4 @@
-package test;
+package test.mapper;
 
 import static org.junit.jupiter.api.Assertions.*;
